requirements.txt
python-telegram-bot==21.6
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
Flask==3.0.3
gunicorn==22.0.0